pseudo-implementation of findOpenContractsNearby and findOpenContractsDuringTimeframe

+++ src/locations/Land.kt
@@ -8,5 +8,5 @@ data class Land(
     // TODO: these need to be ordered in a consistent way (e.g., starting from North and going clockwise)
     val bounds: List<Geolocation>,
     // physical size, as measured in km^2
-    val magnitude: Int
+    val magnitude: Long
 )
